# Conflicts:
#	app/src/main/java/com/waz/zclient/KotlinServices.kt
#	app/src/main/kotlin/com/waz/zclient/feature/backup/BackUpViewModel.kt
#	app/src/main/kotlin/com/waz/zclient/feature/backup/di/BackUpModule.kt
#	app/src/main/kotlin/com/waz/zclient/feature/backup/metadata/MetaDataHandlerDataSource.kt
#	app/src/main/kotlin/com/waz/zclient/feature/backup/usecase/CreateBackUpUseCase.kt
#	app/src/main/scala/com/waz/zclient/preferences/pages/BackupExportView.scala
#	app/src/test/kotlin/com/waz/zclient/feature/backup/BackUpDataSourceTest.kt
#	app/src/test/kotlin/com/waz/zclient/feature/backup/metadata/MetaDataHandlerTest.kt
#	app/src/test/kotlin/com/waz/zclient/feature/backup/usecase/CreateBackUpUseCaseTest.kt

+++ app/src/main/kotlin/com/waz/zclient/shared/user/mapper/UserMapper.kt
@@ -86,7 +86,7 @@ class UserMapper {
         private const val DEFAULT_ACCENT_ID_VALUE = 0
         private val DEFAULT_SKEY_VALUE = String.empty()
         private const val DEFAULT_CONNECTION_VALUE = "unconnected"
-        private const val DEFAULT_CONNECTION_TIMESTAMP_VALUE: Int = 0
+        private const val DEFAULT_CONNECTION_TIMESTAMP_VALUE: Long = 0
         private const val DEFAULT_RELATION_VALUE = "Other"
         private const val DEFAULT_VERIFIED_VALUE = "UNKNOWN"
         private const val DEFAULT_DELETED_VALUE = false
